Include telemetry opt-out instructions in Czech
Remove the "aby zlepšil uživatelské prostředí" (as of now,
the Privacy Policy lists 17 reasons data is collected!).

Include a note that the telemetry may send the content
of your files. Citing the current privacy policy:

> The data we collect can include the following:
> [...]
> **Text, inking, and typing data.** Text, inking, and typing
> data and related information.
> [...]
> **Content.** Content of your files and communications you
> input, upload, receive, create, and control.

That might cover anything you type in or open in the editor.
(Since the telemetry is closed-source, it's hard to tell if
they're actually sending this.)